Added revolver
Revolver is almost fully implemented, got 3/4 animations done (just missing the reload animation) and a placeholder for the gunshot. I implemented the gun firing animation and idle animation, so now the gun shoots when the player left clicks. i didnt get to do the walking animation yet, i wasnt sure how to implement that but also because i need to go to bed for an appointment.