Added full animations to revolver
The camera is a bit finnicky, idk why but it clips parts of the reload animation so it looks like nothing is filling the chambers when the reload happens. I think its the near clipping range since when i move the gun far away, the animation has no missing objects. also added reload sound support with a funny placeholder sound for reloads.