minor bug fixes
Fixed the bug where the gun would become squished when the player crouched. Added a check for if the gun is blocked by a wall.  Added a animation for the gun for if the gun is blocked.

I tried to fix the gun clipping through the wall, but I haven't found a solution yet.